% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deck_quarters}
\alias{deck_quarters}
\title{create a shuffled deck of quarters}
\usage{
deck_quarters(quarters_table = table_quarters(),
  n_cards = sum(unlist(quarters_table)))
}
\arguments{
\item{quarters_table}{information on the quarters available in the game}

\item{n_cards}{integer. number of cards in the deck}
}
\description{
create a shuffled deck of quarters
}

