% Generated by roxygen2 (4.0.2): do not edit by hand
\name{game}
\alias{game}
\title{play a random game of Citadels}
\usage{
game(players = list_players(4), quarters_deck = deck_quarters(),
  n_coins_init = 2, n_cards_init = 4, max_city_size = 8, quiet = TRUE)
}
\arguments{
\item{players}{list of players}

\item{quarters_deck}{list of quarters in the deck}

\item{n_coins_init}{integer. number of coins for each player at the beginning}

\item{n_cards_init}{integer. number of cards for each player at the beginning}

\item{max_city_size}{integer. if any player builds \code{max_city_size} the
game stops at the end of the turn.}

\item{quiet}{logical. disable text output}
}
\description{
play a random game of Citadels
}
\examples{
require(citadels)

game(list_players(3), quiet = FALSE)
}

