% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_players}
\alias{list_players}
\title{create a list of players}
\usage{
list_players(n_players = 2, names = strsplit("ABCDEFGH", "")[[1]])
}
\arguments{
\item{n_players}{integer. number of players in the list}

\item{names}{character vector. names of the players}
}
\description{
create a list of players
}

