% Generated by roxygen2 (4.0.2): do not edit by hand
\name{player}
\alias{player}
\title{create a player}
\usage{
player(name, has_crown = FALSE)
}
\arguments{
\item{name}{string. name of the player}

\item{has_crown}{logical. \code{TRUE} if the player has the crown}
}
\description{
create a player
}

